[TR-210] Move route-linked data from select-view-settings to route-data
