+++ public/canjs/controls/select-report-type/select-report-type.js
@@ -1,7 +1,6 @@
-import { StacheElement, type, ObservableObject, ObservableArray, value } from "../../../can.js";
+import { StacheElement } from "../../../can.js";
 import { DROPDOWN_LABEL } from "../../../shared/style-strings.js";
 
-import {saveJSONToUrl,updateUrlParam} from "../../routing/state-storage.js";
 import routeData from "../../routing/route-data.js";
 
 import "../status-filter.js";
