+++ public/timeline-report.js
@@ -87,9 +87,6 @@ export class TimelineReport extends StacheElement {
               groupBy:to="this.groupBy"
               releasesToShow:to="this.releasesToShow"
               primaryReportBreakdown:to="this.primaryReportBreakdown"
-              primaryReportType:from="this.routeData.primaryReportType"
-              primaryIssueType:from="this.primaryIssueType"
-              secondaryIssueType:from="this.secondaryIssueType"
               statuses:from="this.statuses"
               derivedIssues:from="this.routeData.derivedIssues"
               isLoggedIn:from="this.loginComponent.isLoggedIn"
